+++ packages/app/src/components/catalog/EntityPage/ContextMenuAwareEntityLayout.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode, useMemo, useState } from 'react';
 import { EntityLayout } from '@backstage/plugin-catalog';
 import getMountPointData from '../../../utils/dynamicUI/getMountPointData';
-import { MenuIcon } from '../../Root/Root';
+import { MenuIcon } from '../../Root/MenuIcon';
 import { IconComponent } from '@backstage/core-plugin-api';
 
 const makeIcon =
